+++ frontend/runtime_configs/dev.env
@@ -1,4 +1,6 @@
 RUNTIME_VAR_API_URL=https://api.cellxgene.dev.single-cell.czi.technology
+RUNTIME_VAR_DE_API_URL=https://de-api.cellxgene.dev.single-cell.czi.technology
+RUNTIME_VAR_WMG_API_URL=https://wmg-api.cellxgene.dev.single-cell.czi.technology
 RUNTIME_VAR_CELLGUIDE_DATA_URL=https://cellguide.cellxgene.dev.single-cell.czi.technology
 RUNTIME_VAR_CENSUS_MODELS_DATA_URL=https://contrib.cellxgene.cziscience.com
 RUNTIME_VAR_PLAUSIBLE_DATA_DOMAIN=cellxgene.staging.single-cell.czi.technology
