+++ frontend/runtime_configs/local.env
@@ -1,4 +1,6 @@
 RUNTIME_VAR_API_URL=https://backend.corporanet.local:5000
+RUNTIME_VAR_DE_API_URL=https://backend-de.corporanet.local:5000
+RUNTIME_VAR_WMG_API_URL=https://backend-wmg.corporanet.local:5000
 RUNTIME_VAR_CELLGUIDE_DATA_URL=https://cellguide.cellxgene.dev.single-cell.czi.technology
 RUNTIME_VAR_CENSUS_MODELS_DATA_URL=https://cellxgene-contrib-public.s3.us-west-2.amazonaws.com
 RUNTIME_VAR_PLAUSIBLE_DATA_DOMAIN=cellxgene.staging.single-cell.czi.technology
