+++ frontend/runtime_configs/prod.env
@@ -1,4 +1,6 @@
 RUNTIME_VAR_API_URL=https://api.cellxgene.cziscience.com
+RUNTIME_VAR_DE_API_URL=https://de-api.cellxgene.cziscience.com
+RUNTIME_VAR_WMG_API_URL=https://wmg-api.cellxgene.cziscience.com
 RUNTIME_VAR_CELLGUIDE_DATA_URL=https://cellguide.cellxgene.cziscience.com
 RUNTIME_VAR_CENSUS_MODELS_DATA_URL=https://contrib.cellxgene.cziscience.com
 RUNTIME_VAR_PLAUSIBLE_DATA_DOMAIN=cellxgene.cziscience.com
