+++ frontend/src/configs/build.js
@@ -1,5 +1,7 @@
 const configs = {
   API_URL: "RUNTIME_VAR_API_URL",
+  DE_API_URL: "RUNTIME_VAR_DE_API_URL",
+  WMG_API_URL: "RUNTIME_VAR_WMG_API_URL",
   CELLGUIDE_DATA_URL: "RUNTIME_VAR_CELLGUIDE_DATA_URL",
   PLAUSIBLE_DATA_DOMAIN: "RUNTIME_VAR_PLAUSIBLE_DATA_DOMAIN",
   CENSUS_MODELS_DATA_URL: "RUNTIME_VAR_CENSUS_MODELS_DATA_URL",
