+++ python_dependencies/backend/common/server/requirements.txt
@@ -1,0 +1,8 @@
+connexion[swagger-ui]~=2.14.2
+gunicorn[gevent]~=22.0.0
+furl~=2.1.2
+Flask-Cors~=3.0.6
+flask-server-timing~=0.1.2
+ddtrace~=2.8.5
+python-json-logger
+boto3~=1.34.114
