+++ python_dependencies/backend/wmg/requirements.txt
@@ -1,0 +1,7 @@
+pandas~=2.2.2
+pydantic~=1.10.7
+numba~=0.59.1
+cellxgene_ontology_guide~=1.0.0
+tiledb
+psutil~=5.9.8
+pyarrow==16.0.0
